clang-problem: test ll_traverse_test crashes on clang
Crash happens on ll_traverse. Same code generated with gcc does not have this
problem. If the printf is uncommented the test succeeds with clang also.

Valgrind reports "Conditional jump or move depends on uninitialised value(s)"

--

The problem is that the callback function for ll_traverse does not return a
value. Gcc is apparently more lenient or it places a default value as a return
value, where clang displays only a compile time warning.